+++ src/OpenLoco/src/Ui/Windows/News/News.cpp
@@ -174,7 +174,7 @@ namespace OpenLoco::Ui::Windows::NewsWindow
             }
         }
 
-        static SavedView getView(Window* self, Message* news, uint16_t itemId, MessageItemArgumentType itemType, bool* selectable)
+        static SavedView getView(Window* self, const Message* news, uint16_t itemId, MessageItemArgumentType itemType, bool& selectable)
         {
             SavedView view;
             view.mapX = -1;
@@ -194,7 +194,7 @@ namespace OpenLoco::Ui::Windows::NewsWindow
                     view.surfaceZ = World::TileManager::getHeight({ view.mapX, view.mapY }).landHeight;
                     view.rotation = WindowManager::getCurrentRotation();
                     view.zoomLevel = ZoomLevel::half;
-                    *selectable = true;
+                    selectable = true;
                     break;
                 }
 
@@ -207,7 +207,7 @@ namespace OpenLoco::Ui::Windows::NewsWindow
                     view.surfaceZ = station->z;
                     view.rotation = WindowManager::getCurrentRotation();
                     view.zoomLevel = ZoomLevel::full;
-                    *selectable = true;
+                    selectable = true;
                     break;
                 }
 
@@ -220,7 +220,7 @@ namespace OpenLoco::Ui::Windows::NewsWindow
                     view.surfaceZ = World::TileManager::getHeight({ view.mapX, view.mapY }).landHeight;
                     view.rotation = WindowManager::getCurrentRotation();
                     view.zoomLevel = ZoomLevel::half;
-                    *selectable = true;
+                    selectable = true;
                     break;
                 }
 
@@ -247,7 +247,7 @@ namespace OpenLoco::Ui::Windows::NewsWindow
                     view.flags = (1 << 15);
                     view.zoomLevel = ZoomLevel::full;
                     view.rotation = WindowManager::getCurrentRotation();
-                    *selectable = true;
+                    selectable = true;
                     break;
                 }
 
@@ -256,7 +256,7 @@ namespace OpenLoco::Ui::Windows::NewsWindow
                     // TODO: Do this better
                     view.zoomLevel = enumValue(SubjectType::companyFace);
                     self->invalidate();
-                    *selectable = true;
+                    selectable = true;
                     break;
 
                 case MessageItemArgumentType::location:
@@ -265,7 +265,7 @@ namespace OpenLoco::Ui::Windows::NewsWindow
                     view.surfaceZ = World::TileManager::getHeight({ view.mapX, view.mapY }).landHeight;
                     view.zoomLevel = ZoomLevel::full;
                     view.rotation = WindowManager::getCurrentRotation();
-                    *selectable = true;
+                    selectable = true;
                     break;
 
                 case MessageItemArgumentType::unk6:
@@ -277,14 +277,31 @@ namespace OpenLoco::Ui::Windows::NewsWindow
                     // TODO: Do this better
                     view.zoomLevel = enumValue(SubjectType::vehicleImage);
                     self->invalidate();
-                    *selectable = true;
+                    selectable = true;
                     break;
             }
             return view;
         }
 
-        // TODO: deduplicate with initViewport1
-        static void initViewport0(Window& self)
+        struct ViewportLayout
+        {
+            MessageTypeFlags subjectFlag;
+            WidgetIndex_t viewportWidgetId;
+            WidgetIndex_t buttonWidgetId;
+            Point position;
+            Size fullSize;
+            Size halfSize;
+        };
+
+        static constexpr Size kFullSizeNews = { 351, 42 };
+        static constexpr Size kHalfSizeNews = { 174, 42 };
+
+        static constexpr std::array kViewportLayouts = std::to_array<ViewportLayout>({
+            { MessageTypeFlags::hasFirstItem, Common::widx::viewport1, Common::widx::viewport1Button, Point{ 4, 42 }, kFullSizeNews, kHalfSizeNews },
+            { MessageTypeFlags::hasSecondItem, Common::widx::viewport2, Common::widx::viewport2Button, Point{ 186, 42 }, kHalfSizeNews, kHalfSizeNews },
+        });
+
+        static void initViewport(Window& self, const uint8_t subjectIndex)
         {
             SavedView view;
             view.mapX = -1;
@@ -294,73 +311,58 @@ namespace OpenLoco::Ui::Windows::NewsWindow
             view.zoomLevel = (ZoomLevel)0xFFU;
             view.entityId = EntityId::null;
 
-            auto news = MessageManager::get(MessageManager::getActiveIndex());
+            const auto* news = MessageManager::get(MessageManager::getActiveIndex());
             const auto& mtd = getMessageTypeDescriptor(news->type);
+            const auto& layout = kViewportLayouts[subjectIndex];
 
             bool selectable = false;
-
             if (MessageManager::getActiveIndex() != MessageId::null)
             {
-                if (mtd.hasFlag(MessageTypeFlags::hasFirstItem))
+                if (mtd.hasFlag(layout.subjectFlag))
                 {
-                    auto itemType = mtd.argumentTypes[0];
+                    auto itemType = mtd.argumentTypes[subjectIndex];
 
-                    if (news->itemSubjects[0] != 0xFFFF)
+                    if (news->itemSubjects[subjectIndex] != 0xFFFF)
                     {
-                        view = getView(&self, news, news->itemSubjects[0], itemType, &selectable);
+                        view = getView(&self, news, news->itemSubjects[subjectIndex], itemType, selectable);
                     }
                 }
             }
 
-            self.widgets[Common::widx::viewport1].hidden = true;
-            self.widgets[Common::widx::viewport1Button].hidden = true;
+            auto& viewportWidget = self.widgets[layout.viewportWidgetId];
+            auto& buttonWidget = self.widgets[layout.buttonWidgetId];
 
-            if (!view.isEmpty())
-            {
-                self.widgets[Common::widx::viewport1].hidden = false;
-            }
+            viewportWidget.hidden = view.isEmpty();
+            buttonWidget.hidden = !selectable;
 
-            if (selectable)
-            {
-                self.widgets[Common::widx::viewport1Button].hidden = false;
-            }
+            // Update viewport layout
+            const bool juxtapose = subjectIndex == 0 && mtd.hasFlag(MessageTypeFlags::hasSecondItem);
+            const auto& size = juxtapose ? layout.halfSize : layout.fullSize;
 
-            if (_nState.savedView[0] != view)
+            viewportWidget.left = layout.position.x + 2;
+            viewportWidget.right = layout.position.x + size.width - 4;
+            buttonWidget.left = layout.position.x;
+            buttonWidget.right = layout.position.x + size.width;
+
+            // Update viewport focus
+            if (_nState.savedView[subjectIndex] != view)
             {
-                _nState.savedView[0] = view;
-                self.viewportRemove(0);
+                _nState.savedView[subjectIndex] = view;
+                self.viewportRemove(subjectIndex);
                 self.invalidate();
-
-                self.widgets[Common::widx::viewport1].left = 6;
-                self.widgets[Common::widx::viewport1].right = 353;
-                self.widgets[Common::widx::viewport1Button].left = 4;
-                self.widgets[Common::widx::viewport1Button].right = 355;
-
-                if (mtd.hasFlag(MessageTypeFlags::hasSecondItem))
-                {
-                    self.widgets[Common::widx::viewport1].left = 6;
-                    self.widgets[Common::widx::viewport1].right = 173;
-                    self.widgets[Common::widx::viewport1Button].left = 4;
-                    self.widgets[Common::widx::viewport1Button].right = 175;
-                }
 
                 if (!view.isEmpty())
                 {
-                    int16_t x = self.widgets[Common::widx::viewport1].left + 1;
-                    int16_t y = self.widgets[Common::widx::viewport1].top + 1;
-                    Ui::Point origin = { x, y };
+                    auto origin = viewportWidget.position() + Point{ 1, 1 };
 
-                    uint16_t viewportWidth = self.widgets[Common::widx::viewport1].width();
+                    uint16_t viewportWidth = viewportWidget.width();
                     uint16_t viewportHeight = 62;
                     Ui::Size viewportSize = { viewportWidth, viewportHeight };
 
                     if (mtd.hasFlag(MessageTypeFlags::isGeneralNews))
                     {
-                        x = self.widgets[Common::widx::viewport1].left;
-                        y = self.widgets[Common::widx::viewport1].top;
-                        origin = { x, y };
-
-                        viewportWidth = self.widgets[Common::widx::viewport1].width() + 2;
+                        origin = viewportWidget.position();
+                        viewportWidth = viewportWidget.width() + 2;
                         viewportHeight = 64;
                         viewportSize = { viewportWidth, viewportHeight };
                     }
@@ -373,86 +375,6 @@ namespace OpenLoco::Ui::Windows::NewsWindow
                     {
                         ViewportManager::create(&self, 0, origin, viewportSize, view.zoomLevel, view.getPos());
                     }
-                    self.invalidate();
-                }
-            }
-        }
-
-        // TODO: deduplicate with initViewport0
-        static void initViewport1(Window& self)
-        {
-            SavedView view;
-            view.mapX = -1;
-            view.mapY = -1;
-            view.surfaceZ = -1;
-            view.rotation = -1;
-            view.zoomLevel = (ZoomLevel)0xFFU;
-            view.entityId = EntityId::null;
-
-            auto news = MessageManager::get(MessageManager::getActiveIndex());
-            const auto& mtd = getMessageTypeDescriptor(news->type);
-
-            bool selectable = false;
-
-            if (MessageManager::getActiveIndex() != MessageId::null)
-            {
-                if (mtd.hasFlag(MessageTypeFlags::hasSecondItem))
-                {
-                    auto itemType = mtd.argumentTypes[1];
-
-                    if (news->itemSubjects[1] != 0xFFFF)
-                    {
-                        view = getView(&self, news, news->itemSubjects[1], itemType, &selectable);
-                    }
-                }
-            }
-
-            self.widgets[Common::widx::viewport2].hidden = true;
-            self.widgets[Common::widx::viewport2Button].hidden = true;
-
-            if (!view.isEmpty())
-            {
-                self.widgets[Common::widx::viewport2].hidden = false;
-            }
-
-            if (selectable)
-            {
-                self.widgets[Common::widx::viewport2Button].hidden = false;
-            }
-
-            if (_nState.savedView[1] != view)
-            {
-                _nState.savedView[1] = view;
-                self.viewportRemove(1);
-                self.invalidate();
-
-                self.widgets[Common::widx::viewport2].left = 186;
-                self.widgets[Common::widx::viewport2].right = 353;
-                self.widgets[Common::widx::viewport2Button].left = 184;
-                self.widgets[Common::widx::viewport2Button].right = 355;
-
-                if (!view.isEmpty())
-                {
-                    int16_t x = self.widgets[Common::widx::viewport2].left + 1;
-                    int16_t y = self.widgets[Common::widx::viewport2].top + 1;
-                    Ui::Point origin = { x, y };
-
-                    uint16_t viewportWidth = self.widgets[Common::widx::viewport2].width();
-                    uint16_t viewportHeight = 62;
-                    Ui::Size viewportSize = { viewportWidth, viewportHeight };
-
-                    if (mtd.hasFlag(MessageTypeFlags::isGeneralNews))
-                    {
-                        x = self.widgets[Common::widx::viewport2].left;
-                        y = self.widgets[Common::widx::viewport2].top;
-                        origin = { x, y };
-
-                        viewportWidth = self.widgets[Common::widx::viewport2].width() + 2;
-                        viewportHeight = 64;
-                        viewportSize = { viewportWidth, viewportHeight };
-                    }
-
-                    ViewportManager::create(&self, 1, origin, viewportSize, view.zoomLevel, view.getPos());
 
                     self.invalidate();
                 }
@@ -462,8 +384,8 @@ namespace OpenLoco::Ui::Windows::NewsWindow
         // 0x00429209
         void initViewports(Window& self)
         {
-            initViewport0(self);
-            initViewport1(self);
+            initViewport(self, 0);
+            initViewport(self, 1);
         }
 
         // 0x0042A136
@@ -826,29 +748,25 @@ namespace OpenLoco::Ui::Windows::NewsWindow
                 drawStationNews(self, drawingCtx, news);
             }
 
-            if (mtd.hasFlag(MessageTypeFlags::hasFirstItem))
+            if (mtd.hasFlag(MessageTypeFlags::hasFirstItem) && news->itemSubjects[0] != 0xFFFF)
             {
-                if (news->itemSubjects[0] != 0xFFFF)
-                {
-                    auto x = (self.widgets[Common::widx::viewport1Button].left + self.widgets[Common::widx::viewport1Button].right) / 2;
-                    auto y = self.widgets[Common::widx::viewport1Button].bottom - 7;
-                    auto width = self.widgets[Common::widx::viewport1Button].width() - 1;
-                    auto point = Point(x, y);
+                auto& widget = self.widgets[Common::widx::viewport1Button];
+                auto x = widget.midX();
+                auto y = widget.bottom - 7;
+                auto width = widget.width() - 1;
+                auto point = Point(x, y);
 
-                    drawViewportString(drawingCtx, point, width, mtd.argumentTypes[0], news->itemSubjects[0]);
-                }
+                drawViewportString(drawingCtx, point, width, mtd.argumentTypes[0], news->itemSubjects[0]);
             }
-            if (mtd.hasFlag(MessageTypeFlags::hasSecondItem))
+            if (mtd.hasFlag(MessageTypeFlags::hasSecondItem) && news->itemSubjects[1] != 0xFFFF)
             {
-                if (news->itemSubjects[1] != 0xFFFF)
-                {
-                    auto x = (self.widgets[Common::widx::viewport2Button].left + self.widgets[Common::widx::viewport2Button].right) / 2;
-                    auto y = self.widgets[Common::widx::viewport2Button].bottom - 7;
-                    auto width = self.widgets[Common::widx::viewport2Button].width() - 1;
-                    auto point = Point(x, y);
+                auto& widget = self.widgets[Common::widx::viewport2Button];
+                auto x = widget.midX();
+                auto y = widget.bottom - 7;
+                auto width = widget.width() - 1;
+                auto point = Point(x, y);
 
-                    drawViewportString(drawingCtx, point, width, mtd.argumentTypes[1], news->itemSubjects[1]);
-                }
+                drawViewportString(drawingCtx, point, width, mtd.argumentTypes[1], news->itemSubjects[1]);
             }
         }